Move `ManualTextInput` to a separate class
