+++ osu.Framework.Tests/Visual/UserInterface/TestSceneTextBoxEvents.cs
@@ -12,6 +12,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Input;
 using osu.Framework.Testing;
+using osu.Framework.Testing.Input;
 using osuTK;
 using osuTK.Input;
 
@@ -20,7 +21,7 @@ namespace osu.Framework.Tests.Visual.UserInterface
     public partial class TestSceneTextBoxEvents : ManualInputManagerTestScene
     {
         private EventQueuesTextBox textBox;
-        private ManualTextInput textInput;
+        private ManualTextInputSource textInput;
         private ManualTextInputContainer textInputContainer;
 
         private const string default_text = "some default text";
@@ -618,57 +619,12 @@ namespace osu.Framework.Tests.Visual.UserInterface
         public partial class ManualTextInputContainer : Container
         {
             [Cached(typeof(TextInputSource))]
-            public readonly ManualTextInput TextInput;
+            public readonly ManualTextInputSource TextInput;
 
             public ManualTextInputContainer()
             {
                 RelativeSizeAxes = Axes.Both;
-                TextInput = new ManualTextInput();
-            }
-        }
-
-        public class ManualTextInput : TextInputSource
-        {
-            public void Text(string text) => TriggerTextInput(text);
-
-            public new void TriggerImeComposition(string text, int start, int length)
-            {
-                base.TriggerImeComposition(text, start, length);
-            }
-
-            public new void TriggerImeResult(string text)
-            {
-                base.TriggerImeResult(text);
-            }
-
-            public override void ResetIme()
-            {
-                base.ResetIme();
-
-                // this call will be somewhat delayed in a real world scenario, but let's run it immediately for simplicity.
-                base.TriggerImeComposition(string.Empty, 0, 0);
-            }
-
-            public readonly Queue<bool> ActivationQueue = new Queue<bool>();
-            public readonly Queue<bool> EnsureActivatedQueue = new Queue<bool>();
-            public readonly Queue<bool> DeactivationQueue = new Queue<bool>();
-
-            protected override void ActivateTextInput(bool allowIme)
-            {
-                base.ActivateTextInput(allowIme);
-                ActivationQueue.Enqueue(allowIme);
-            }
-
-            protected override void EnsureTextInputActivated(bool allowIme)
-            {
-                base.EnsureTextInputActivated(allowIme);
-                EnsureActivatedQueue.Enqueue(allowIme);
-            }
-
-            protected override void DeactivateTextInput()
-            {
-                base.DeactivateTextInput();
-                DeactivationQueue.Enqueue(true);
+                TextInput = new ManualTextInputSource();
             }
         }
